Update: rewrite popoverStyles as ArrowStyles component
